+++ src/global.h
@@ -164,7 +164,7 @@ inline QString aboutHtml()
     info += "<p>Thanks to all encouraged me writing the application.</p>";
     info += "<p>Built-in Icon Theme taken from Oxygens Icon : <a href='http://www.oxygen-icons.org'>http://www.oxygen-icons.org</a></p>";
     info += "<p>SupraFit website on GitHub: <a href='https://github.com/conradhuebler/SupraFit'>https://github.com/conradhuebler/SupraFit</a></p>";
-    info += "<p>If you obtain results with SupraFit, I kindly ask to cite: </p><p>C. Hübler, conradhuebler/SupraFit: Version 1.7.0 2019, Zenodo. <a href='http://doi.org/10.5281/zenodo.3364570'>http://doi.org/10.5281/zenodo.3364570</a>.<p>";
+    info += "<p>If you obtain results with SupraFit, I kindly ask to cite: </p><p>C. Hübler, conradhuebler/SupraFit: 2019, Zenodo. <a href='http://doi.org/10.5281/zenodo.3364569'>http://doi.org/10.5281/zenodo.3364569</a>.<p>";
     info += "<p>SupraFit has been compilied on " + QString::fromStdString(__DATE__) + " at " + QString::fromStdString(__TIME__) + ".\n";
 #ifdef noto_font
     info += "<p>SupraFit uses and provides some selected Google Noto Font, see <a href='https://github.com/googlei18n/noto-fonts'>https://github.com/googlei18n/noto-fonts</a></p>";
@@ -185,7 +185,7 @@ inline QString about()
     info += "\tSpecial thanks to Alexander König for testing the Quickstart, helpful comments and feedback.\n\n\n";
     info += "\tThanks to all encouraged me writing the application.\n\n";
     info += "\tSupraFit website on GitHub: https://github.com/conradhuebler/SupraFit\n\n";
-    info += "\tIf you obtain results with SupraFit, I kindly ask to cite: C. Hübler, conradhuebler/SupraFit: Version 1.7.0 2019, Zenodo. http://doi.org/10.5281/zenodo.3364570\n";
+    info += "\tIf you obtain results with SupraFit, I kindly ask to cite: C. Hübler, conradhuebler/SupraFit: 2019, Zenodo. http://doi.org/10.5281/zenodo.3364569\n";
     info += "\tSupraFit has been compilied on " + QString::fromStdString(__DATE__) + " at " + QString::fromStdString(__TIME__) + ".\n\n";
     info += "\t*********************************************************************************************************\n\n";
     return info;