+++ src/ui/mainwindow/modelwidget.h
@@ -78,7 +78,7 @@ class ModelWidget : public QWidget {
     Q_OBJECT
 
 public:
-    ModelWidget(QSharedPointer<AbstractModel> model, Charts charts, bool readonly = false, QWidget* parent = 0);
+    ModelWidget(QSharedPointer<AbstractModel> model, Charts charts, bool readonly = false, QWidget* parent = NULL);
     virtual ~ModelWidget() override;
     virtual inline QSize sizeHint() const override { return QSize(125, 25); }
     QSharedPointer<AbstractModel> Model() const { return m_model; }
@@ -124,7 +124,6 @@ private:
     QVBoxLayout* m_layout;
     bool m_pending;
     QList<int> ActiveSignals();
-    void DiscreteUI();
     void resizeButtons();
     void CollectParameters();
     void Data2Text();